ENG-12809: Go driver doesn't read some responses in async mode
After deep investigation, I have come into conclusion that the responses arent processed fast enough resulting into the the response channel to block .

When the response channel is full, the  voltdb instances closes the connection when  responses are not read for a specific timeout limit leading to broken pipe error and connection reset error.

This is  workaround based from observation of the voter benchmark. I was doing increments of multiples of 10 to find the right buffer size.