Fix duration flag for voter benchmrark
The duration flag `--duration` is  supposed to limit how long the
benchmark will run. This was not properly desinged, and didint was as it
was supposed to.

I needed a way to shorten the benchmarks to take profile snapshot for
debugging.

This Commit fix that by using the context API